# src/main/java/com/classroom/core/dto/auth/LoginRequest.java
package com.classroom.core.dto.auth;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {

    @NotBlank
    private String username;

    @NotBlank
    private String password;
}