# src/main/java/com/classroom/core/dto/comment/CreateCommentRequest.java
package com.classroom.core.dto.comment;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class CreateCommentRequest {

    @NotBlank
    @Size(min = 1, max = 5000)
    private String text;
}